Refactor Carousel component to use custom hook for animation
- Migrate animation logic from Carousel to a new useCarouselAnimation hook
